Created a prototype (4 files). Removed the echo server, which is obsolete.
ee-client is the core module. It controls communication with the Engine,
and dispatches incoming messages to the appropriate functions.

ee-utils are contains non-standard emacs utilities.

ee-log is just for logging = debugging.

ee-server is just for testing. It plays the role of the Engine. It contains
methods to send all kinds of Engine messages.